pyproject.toml
[project]
name = "wanted"
version = "0.1.0"
description = ""
authors = [
    {name = "YisooKim",email = "gashe7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pytest-asyncio = "^0.26.0"
greenlet = "^3.1.1"
cryptography = "^44.0.2"
httpx = "^0.28.1"


[tool.mypy]
python_version = "3.13"
exclude = ['^tests/']
strict = true
ignore_missing_imports = true
disable_error_code = ["misc"]

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true


[tool.pytest.ini_options]
addopts = "-ra -q"
asyncio_mode = "auto"
testpaths = [
    "test",
]
